minor fixes to metrics
