+++ front/components/home/content/Product/IntroSection.tsx
@@ -1,13 +1,27 @@
 import { Div3D, Button, Hover3D, RocketIcon } from "@dust-tt/sparkle";
 import Link from "next/link";
 import React from "react";
-
-import { ImgBlock } from "@app/components/home/ContentBlocks";
+import { ImgBlock, MetricSection } from "@app/components/home/ContentBlocks";
 
 import { H1, P, Strong } from "@app/components/home/ContentComponents";
 import TrustedBy from "@app/components/home/TrustedBy";
 import { classNames } from "@app/lib/utils";
 import { ValuePropSection } from "@app/components/home/content/Product/ValuePropSection";
+
+const METRICS = {
+  metrics: [
+    {
+      value: "90%",
+      description: <>Weekly active users at Alan</>,
+    },
+    {
+      value: "50,000",
+      description: <>Annual hours saved at Qonto</>,
+    },
+  ],
+  from: "from-amber-200",
+  to: "to-amber-500",
+};
 
 const VideoPlayer = () => {
   return (
@@ -68,6 +82,7 @@ export function IntroSection() {
         </div>
         <TrustedBy />
         <ValuePropSection />
+        <MetricSection {...METRICS}></MetricSection>
       </div>
     </div>
   );
